refactor(combat): split damage/threat along QR-roll/pipeline and aggro/player-combat seams
Both files were over the 700-line hard cap (CLAUDE.md file organization).
Pure code movement — no logic, signature, or behavior changes.

damage.rs (839) -> damage/:
  - qr.rs        QR scoring + beta-distribution roll -> result code
  - pipeline.rs  multi-stage damage-application formula + absorption shields
  - mod.rs       re-exports calculate_qr/calculate_result/qr_rand_to_result_code/
                 QrResult/calculate_damage (public surface unchanged)

threat.rs (733) -> threat/:
  - aggro.rs          NPC threat-table mutation, Idle->Fighting preempt,
                      leash/attack-range constants, generate_threat
  - player_combat.rs  threatened_mobs / BSF_IN_COMBAT bookkeeping +
                      dead-NPC sweep
  - mod.rs            re-exports all prior public items

Inline test modules moved alongside the functions they exercise; test
bodies are byte-identical (import paths fixed only).

Part of #529.
